+++ client/components/dashboard/monthlySummary/monthlySummary.spec.js
@@ -16,7 +16,9 @@ describe('<MonthlySummary />', () => {
     orderActions: {
       getMonthlySummary() { }
     },
-    loading: false
+    loading: false,
+    currentMonthSales: 60.34,
+    monthAverage: 480.33
   };
 
   it('should set the title heading to "Monthly Sales Volume"', () => {
@@ -155,5 +157,167 @@ describe('<MonthlySummary />', () => {
     expect(actual).equals(expected);
 
   });
+
+  it('should set the average monthly amount to the average across all orders when mapping from the monthlySummary', () => {
+
+    const expected = 60.38;
+    const state = {
+      loading: {
+        loadingMonthlySummary: false
+      },
+      orders: {
+        monthlySummary: [{
+          monthDisplayName: 'Jan',
+          total: 40.38,
+          orderCount: 4
+        },
+        {
+          monthDisplayName: 'Feb',
+          total: 80.38,
+          orderCount: 8
+        }]
+      }
+    };
+
+    const actual = mapStateToProps(state)
+      .monthAverage;
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should set the monthlyAverage to 0 if no orders exist', () => {
+
+    const expected = 0;
+    const state = {
+      loading: {
+        loadingMonthlySummary: false
+      },
+      orders: {
+        monthlySummary: []
+      }
+    };
+
+    const actual = mapStateToProps(state)
+      .monthAverage;
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should set the current month sales amount accordingly', () => {
+
+    const expected = 40.38;
+    const state = {
+      loading: {
+        loadingMonthlySummary: false
+      },
+      orders: {
+        monthlySummary: [{
+          monthDisplayName: 'Jan',
+          total: 40.38,
+          orderCount: 4
+        },
+        {
+          monthDisplayName: 'Feb',
+          total: 80.38,
+          orderCount: 8
+        }]
+      }
+    };
+
+    const actual = mapStateToProps(state)
+      .currentMonthSales;
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should set the current month sales amount to 0 if there are no sales in the store', () => {
+
+    const expected = 0;
+    const state = {
+      loading: {
+        loadingMonthlySummary: false
+      },
+      orders: {
+        monthlySummary: []
+      }
+    };
+
+    const actual = mapStateToProps(state)
+      .currentMonthSales;
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should set the sales progress bar percentage accordingly', () => {
+
+    const expected = 13;
+
+    const instance = shallow(<MonthlySummary.WrappedComponent {...props} />)
+      .instance();
+
+    const actual = instance.getSalesProgressBarPercentage();
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should set the sales progress bar percentage to 0 if the value is NaN', () => {
+
+    const expected = 0;
+
+    const updatedProps = {
+      ...props,
+      monthAverage: 'foobar'
+    };
+
+    const instance = shallow(<MonthlySummary.WrappedComponent {...updatedProps} />)
+      .instance();
+
+    const actual = instance.getSalesProgressBarPercentage();
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should set the sales progress bar percentage to 100 if the value exceeds the monthly average', () => {
+
+    const expected = 100;
+
+    const updatedProps = {
+      ...props,
+      currentMonthSales: 100,
+      monthAverage: 30
+    };
+
+    const instance = shallow(<MonthlySummary.WrappedComponent {...updatedProps} />)
+      .instance();
+
+    const actual = instance.getSalesProgressBarPercentage();
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should set the sales progress bar percentage as the value', () => {
+
+    const expected = 13;
+
+    const wrapper = shallow(<MonthlySummary.WrappedComponent {...props} />)
+      .find(ProgressBar);
+
+
+    const actual = wrapper
+      .at(0)
+      .props()
+      .value;
+
+    expect(actual).equals(expected);
+
+  });
+
 
 });