+++ client/utilities/ordersUtility.js
@@ -23,7 +23,7 @@ export function getOrderStatusDescription(orderStatusId) {
 }
 
 export function mapOrderSummary(orderData, dateString) {
-  
+
   const date = new Date(dateString);
   const monthValue = date.getMonth();
   const year = date.getFullYear();
@@ -53,4 +53,23 @@ export function mapOrderSummary(orderData, dateString) {
     });
 
   return monthSummary;
+}
+
+export function monthlyAverage(monthlySummary) {
+
+  const size = monthlySummary.length;
+  
+  if (!size) {
+    return 0;
+  }
+
+  const totalSales = monthlySummary
+    .reduce((previous, next) => {
+      return previous + next.total;
+    }, 0);
+
+  const monthlyAverageValue = totalSales / size;
+  const average = Number(monthlyAverageValue.toFixed(2));
+  
+  return average;
 }